+++ lib/constants/app_colors.dart
@@ -4,4 +4,5 @@ class AppColors {
   static final Color greyBackground = Color(0xfff5f5f5);
   static final Color secondary = Color(0xff3B3E5B);
   static final Color inactiveBlack = Color.fromRGBO(124, 126, 146, 0.56);
+  static final Color greenColor = Color.fromRGBO(76, 175, 80, 1);
 }
