+++ pubspec.yaml
@@ -9,19 +9,18 @@ environment:
   sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
+  cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
-
-  cupertino_icons: ^1.0.2
-
   flutter_svg: ^0.23.0+1
+  intl: ^0.17.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
+  dart_code_metrics: ^4.6.0
 
   flutter_lints: ^1.0.0
-  dart_code_metrics: ^4.6.0
+  flutter_test:
+    sdk: flutter
 
 flutter:
